v1alpha5: rename Component.Component to Component.Path (#296)
Mike suggested Component.Path is more clear, I agree.  I'm not changing
the `holos_component` tag name though because that would break a bunch
of existing code out there.

This field is used fairly pervasively, so it would be an impactful
backwards incompatible change if we changed v1alpha4.  We establish
v1alpha5 and make the change there instead.

It's a lot of copying and pasting to get a new core API version, but
it's actually fairly mechanical and quick.