v1alpha5: add Aspects fields to author api resources
Mike studied v1alpha4 and concluded he'd need to fork the Cluster and
Fleets definitions to add an OS field to the cluster.  He needs to pass
a value of the cluster OS to components.

The Author API is meant to be composeable, so it's a problem this isn't
well supported and obvious.

This patch defines Aspects as a type, which is an open struct in CUE for
exclusive use by end users.  The result is Mike can define an os field
within Cluster.Aspects and that will be available when rendering a
component.

For example, in `myschema.cue`

    package holos

    import author "github.com/holos-run/holos/api/author/v1alpha5"

    #Cluster: author.#Cluster & {
      Aspects: OS: string
    }

    _Fleets: author.#StandardFleets & {
      workload: local: #Cluster & {
        Aspects: OS: "Linux"
      }
    }

Note this is probably overkill, the same use case could be accomplished
by using the cluster name as a lookup key in a user defined _Clusters
struct.  For example `os: _Clusters[_Tags.component.cluster].os` where
`_Clusters: [Name=string]: #Cluster` and `#Cluster: os: string`.