concepts: add sequence diagram for platform rendering (#298)
It took Mike a weekend of pouring over the docs to understand how it
works and he came up with this sequence diagram.  It's correct and a
good high level start, definitely better than what we have in the
concepts page.

This is an improvement, but isn't sufficient to solve the problem that
it took him too long to get an understanding.  While I feel it's a win
he was able to without asking many questions, it's clear the best
information is currently in the Core API docs and we need to lift this
up and summarize it better in the Concepts page, which hasn't been
updated since v1alpha2.

Follow this patch up with a more detailed sequence diagram of component
rendering and how the artifacts are generated and transformed.

Make pass through the concepts and summarize the detailed information
described in the v1alpha4/v1alpha5 core api docs.