define #KargoProjectBuilder and podinfo project
This patch defines a #KargoProjectBuilder schema which takes a
collection of components and stages as input, then produces a Project as
output.  The project is a collection of components that vary by the
Stage suitable for use with a Kargo promotion process.

The KargoProjectBuilder definition doesn't define the Kargo Project and
Stages yet, but it creates a clear place to do so.