manage the kargo project namespace
This patch adds a kargo-project component to the `#KargoProjectBuilder`
definition.  This component manages the Kargo Project resource, which is
effectively a Namespace with the same name as the holos project and the
kargo project.

The project builder adds a similar Namespace resource with the same name
and labeled appropriately for adoption by the Kargo Project.  The reason
is to allow customization and management of the Namespace without having
Kargo manage the lifecycle.

This patch also moves the component output directory to a project
namespace.  This is to avoid collisions when multiple projects are
managed with the `#KargoProjectBuilder` definition.