manage kargo project promotion policies
This patch adds a KargoProjects top level field similar to the holos
Projects top level field.  This structure is useful to store attributes
for use by Kargo itself.  The first use case is to define automatic
promotion policies for nonprod stages.  We'll use the `#KargoProject`
definition to define how code is promoted through the stages later.