manage kargo stages and warehouse for podinfo
This patch manages the Kargo Stage and Warehouse resource for the
podinfo Project.  The requested freight sequence is statically defined
as a structure in the PODINFO KargoProject configured in
projects.podinfo.cue.

This isn't ideal, ideally each stage would define a reference to a
previous stage of where to request freight from, but it would be overly
complex.  For now it makes sense to statically define it.   We at least
limit the static configuration to the nonprod stages.  All prod stages
are dynamically configured and request freight from the uat stage.

Note the stages aren't fully functional yet, we haven't written the
kustomization.yaml stub file for Kargo to edit.  We'll do that in a
follow up commit.