move kargo stage configuration to #Stages
Previously a separate struct was used to configure the stage progression
using the requestedFreight field.  This is a problem because it makes
sense to define the progression as an aspect of the #Stage definition.

This patch fixes the problem by composing additional fields into the
`#Stage` definition when Kargo is built into the platform configuration,
which it is by default in this repository.

Using the prior stage name and the special string "direct" the
KargoProjectBuilder definition constructs the promotions field for use
by the kargo-stages component.