fix podinfo httproute backend ref
Previously the HTTPRoute for podinfo had an entry for the podinfo
namespace.  This is a problem because that's the Kargo project
namespace, no service is running there so we were getting 500 errors.

This patch fixes the problem by removing the entry.  We switch to an
approach of registering each stages backend namespace with the holos
project httpRoutes field.  The kargo project builder accepts a backend
ref as a template to compose so it doesn't need to concern itself with
the podinfo service.