refactor holos and kargo projects
Previously the KargoProjectBuilder was a bit messy.  Everything is
defined at the root, but these definitions don't heavily use tags like
the #ComponentConfig definition does, so it makes sense to organize them
better into packages.

There's also the problem of "which project?" when speaking of a project.
There's a holos project definition for managing a group of related holos
components, and there's also a kargo project definition for managing the
kargo promotion pipeline.  These two concepts are closely related but
not the same.

This patch splits the schema definitions into two packages, one for
kargo and one for holos.  The holos package name is already used, so we
place it in the v1alpha5 package to avoid an import cycle.

With this patch the output of `holos render platform` is identical to
the previous commit, indicating the refactor is on the right track.
Another commit is necessary to render each artifact from the buildplans.