manage cert-manager using external data - needs holos 0.101.7
We need a way for Kargo to edit a yaml file so it can promote
cert-manager.  We'll use the yaml-update step to update the helm chart
version, then execute holos render platform from a ci job that runs only
when changes to the data/** directory are detected.

This patch needs Holos 0.101.7 or later to work.  The updated type
definitions are available with holos init platform v1alpha5