manage cert-manager kargo promotion
This patch implements the promotion steps and has been tested up through
triggering the ArgoCD sync.

Note the `git-wait-for-pr` step takes quite a while to complete.  It
appears to poll on a 5 minute interval.

After playing with this a bit it seems we may want to stop the Kargo
process at opening the PR.