docs: flatten a complex hierarchy
Migrating the helm hierarchy from the original example wasn't a great
example because the hierarchy was simple and there were few overrides.

This change goes further with a second example that generates a complex
hierarchy with 11 layers of overrides and data placed randomly through
the levels.   The document shows how holos helps flatten the hierarchy
into one layer to simplify the configuration.