api: define BuildContext to bind tempDir
Previously, the build plan CUE code has no way to refer to the temporary
directory managed and owned by the `holod render component` command.
This patch adds a placeholder context field for holos to fill in just
before exporting the final build plan from CUE.

This allows the user to refer to the tempDir field without the value
being concrete, useful for constructing an argument vector for the new
Command task.